count values as constructor parameter

+++ app/src/main/java/com/example/haushaltsbuch/data/model/gamification/Coupon.kt
@@ -1,12 +1,13 @@
 package com.example.haushaltsbuch.data.model.gamification
 
 import com.example.haushaltsbuch.data.model.persons.Person
+import java.math.BigDecimal
 import java.util.*
 
 class Coupon(
         val person: Person,
         val code: String,
-        val minimumOrderValue: Int,
+        val minimumOrderValue: BigDecimal,
         val expirationDate: Date) {
 }
