+++ app/src/main/java/com/example/haushaltsbuch/data/model/mediation/PriceModel.kt
@@ -1,10 +1,10 @@
 package com.example.haushaltsbuch.data.model.mediation
 
 import com.example.haushaltsbuch.data.model.persons.BillingInterval
-import java.util.*
+import java.math.BigDecimal
 
 class PriceModel(
-        val price: Int,
+        val price: BigDecimal,
         val isSubscription: Boolean,
         val interval: BillingInterval?,
         ) {
